pyproject.toml
[project]
name = "agent-web-qa-bot"
version = "1.0.0"
description = "김동휘 웹사이트 전용 질의응답 챗봇 - A2A 프로토콜 기반 지능형 QA 에이전트"
authors = [
    {name = "Kim Donghwi", email = "dev@example.com"}
]
requires-python = ">=3.12"
dependencies = [
    # A2A Protocol
    "a2a-sdk[http-server]>=0.3.0",
    # AI/LLM Integration
    "google-genai>=1.32.0",
    "openai>=1.0.0",
    # Web Server
    "uvicorn>=0.34.2",
    "starlette>=0.36.0",
    # HTTP Client for MCP
    "aiohttp>=3.9.0",
    "httpx>=0.24.0",
    # Configuration
    "python-dotenv>=1.0.0",
    # MCP Client
    "mcp>=1.0.0",
    # Utilities
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
agent = "src.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py312']

[tool.ruff]
line-length = 88
target-version = "py312"